+++ src/test/java/ordertests/CreateOrderTest.java
@@ -1,4 +1,4 @@
-package order.tests;
+package ordertests;
 
 import io.qameta.allure.Description;
 import io.qameta.allure.junit4.DisplayName;
