+++ src/test/java/ordertests/GetOrderTest.java
@@ -1,4 +1,4 @@
-package order.tests;
+package ordertests;
 
 import io.qameta.allure.Description;
 import io.qameta.allure.junit4.DisplayName;
