+++ src/test/java/usertests/CommonUserTest.java
@@ -1,28 +1,19 @@
-package order.tests;
+package usertests;
 
-import models.OrderModel;
 import models.UserModel;
 import org.junit.After;
 import org.junit.Before;
-import steps.IngredientStep;
-import steps.OrderSteps;
 import steps.UserSteps;
 
-public class CommonOrderTest {
+public class CommonUserTest {
 
-    protected static OrderModel order;
-    protected static OrderSteps orderSteps;
     protected static UserSteps userSteps;
     protected static UserModel user;
-    protected static IngredientStep ingredientStep;
 
     @Before
     public void setUp(){
-        order = new OrderModel();
-        orderSteps = new OrderSteps();
         user = new UserModel();
         userSteps = new UserSteps();
-        ingredientStep = new IngredientStep();
     }
 
     @After
