+++ src/test/java/usertests/CreateUserTest.java
@@ -1,4 +1,4 @@
-package user.tests;
+package usertests;
 
 import io.qameta.allure.Description;
 import io.qameta.allure.junit4.DisplayName;
